enhancement(editor): Fixed isChanged logic, minor refactoring
