+++ src/pages/EditorPage/EditorPanel/helper/entityHelpers.ts
@@ -1,9 +1,117 @@
+import { isEqual } from 'lodash'
+import getFieldInObject from '@helpers/getFieldInObject'
 import { $Any } from '@types'
 
 type Attrib = {
   type?: 'datetime' | 'integer' | 'float'
 }
 
+const getInitialForm = (
+  singleSelect: boolean,
+  {
+    types,
+    statusValues,
+    nameValues,
+    labelValues,
+    assigneesValues,
+    tagValues,
+  }: {
+    types: $Any
+    statusValues: $Any
+    nameValues: $Any
+    labelValues: $Any
+    assigneesValues: $Any
+    tagValues: $Any
+  },
+) => {
+  const disableMessage = 'Names Can Not Be The Same...'
+
+  return {
+    _label: {
+      changeKey: '_label',
+      label: 'Label',
+      field: 'label',
+      type: 'string',
+      disabled: !singleSelect,
+      placeholder: !singleSelect ? disableMessage : '',
+      attrib: {
+        type: 'string',
+      },
+      ...labelValues,
+      value: singleSelect ? labelValues.value : disableMessage,
+    },
+    _name: {
+      changeKey: '_name',
+      label: 'Name',
+      field: 'name',
+      type: 'string',
+      disabled: !singleSelect,
+      placeholder: !singleSelect ? disableMessage : '',
+      attrib: {
+        type: 'string',
+      },
+      ...nameValues,
+      value: singleSelect ? nameValues.value : disableMessage,
+    },
+    _status: {
+      changeKey: '_status',
+      label: 'Status',
+      field: 'status',
+      placeholder: `Mixed (${
+        statusValues.multipleValues && statusValues.multipleValues.join(', ')
+      })`,
+      ...statusValues,
+    },
+    _assignees: {
+      changeKey: '_assignees',
+      label: 'Assignees',
+      field: 'assignees',
+      disabled: !types.includes('task'),
+      placeholder: `Folders Can Not Have Assignees...`,
+      ...assigneesValues,
+    },
+    _tags: {
+      changeKey: '_tags',
+      label: 'Tags',
+      field: 'tags',
+      ...tagValues,
+    },
+  }
+}
+
+const hasChanged = (
+  oldValueObj: $Any,
+  field: string,
+  changeKey: string,
+  newValue: string | null,
+  { nodes, nodeIds }: { nodes: $Any; nodeIds: string[] },
+) => {
+  if (!oldValueObj?.multipleValues && !oldValueObj?.__new) {
+    for (const id of nodeIds) {
+      const ogValue = getFieldInObject(field, nodes[id]?.data)
+
+      // if value undefined or it's a new node skip
+      // (always changed)
+      if (!ogValue || nodes[id]?.__new) {
+        return true
+      }
+
+      // diff value or multipleValues
+      const ownValue = nodes[id]?.data.ownAttrib.includes(changeKey)
+
+      const isSame =
+        (!ownValue && (newValue === undefined || newValue === null || isEqual(newValue, []))) ||
+        (ownValue && isEqual(ogValue, newValue))
+
+      // stop looping if isChanged is ever true
+      if (!isSame) {
+        return true
+      }
+    }
+  }
+
+  return false
+}
 
 const getParentValue = (
   nodes: $Any,
@@ -61,4 +169,4 @@ const getInputProps = (attrib: Attrib) => {
   return props
 }
 
-export { getParentValue, getInputProps, getTypes }
+export { getInitialForm, getParentValue, getInputProps, getTypes, hasChanged }
